requirements.txt
sys
os
csv
numpy
pandas
pydot
graphviz
matplotlib.pyplot
subprocess
PyQt5.QtCore
PyQt5.QtWidgets
PyQt5.QtWebEngineWidgets
matplotlib.backends.backend_qt5agg
matplotlib.figure
pygraphviz
networkx
ddcal
plotly
dash
flask